# Improve performance when creating progressbar renderers