--- a/widget/progressbar_test.go
+++ b/widget/progressbar_test.go
@@ -10,6 +10,21 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+var globalProgressRenderer fyne.WidgetRenderer
+
+func BenchmarkProgressbar(b *testing.B) {
+	var renderer fyne.WidgetRenderer
+	widget := &ProgressBar{}
+	b.ReportAllocs()
+
+	for i := 0; i < b.N; i++ {
+		renderer = widget.CreateRenderer()
+	}
+
+	// Avoid having the value optimized out by the compiler.
+	globalProgressRenderer = renderer
+}
+
 func TestNewProgressBarWithData(t *testing.T) {
 	val := binding.NewFloat()
 	val.Set(0.4)