--- a/widget/progressbarinfinite_test.go
+++ b/widget/progressbarinfinite_test.go
@@ -10,6 +10,21 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+var globalProgressInfRenderer fyne.WidgetRenderer
+
+func BenchmarkProgressbarInf(b *testing.B) {
+	var renderer fyne.WidgetRenderer
+	widget := &ProgressBarInfinite{}
+	b.ReportAllocs()
+
+	for i := 0; i < b.N; i++ {
+		renderer = widget.CreateRenderer()
+	}
+
+	// Avoid having the value optimized out by the compiler.
+	globalProgressInfRenderer = renderer
+}
+
 func TestProgressBarInfinite_Creation(t *testing.T) {
 	bar := NewProgressBarInfinite()
 	// ticker should start automatically