<!-- If this is your first pull request for Fyne please read the contributor docs at:
https://github.com/fyne-io/fyne/wiki/Contributing.
Be sure that your work is based off `develop` branch. --> 

### Description:
<!-- A summary of the change included and which issue it addresses.
Please include any relevant motivation and background. -->

Around a 30+% performance improvement generally. Caches some icon values and allocates the canvas objects inline into the outer struct instead of allocating them in multiple places. We cal likely apply similar changes in many more places and avoid allocating the struct in one place and then allocate each canvas object in a different place. Less for the GC to handle and we run less of a risk for nil pointer de-references.

Benchstat output over 10 runs:
```
goos: linux
goarch: amd64
pkg: fyne.io/fyne/v2/widget
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
                 │   old.txt   │               new.txt               │
                 │   sec/op    │   sec/op     vs base                │
Progressbar-8      491.7n ± 5%   287.4n ± 3%  -41.55% (p=0.000 n=10)
ProgressbarInf-8   491.8n ± 6%   392.5n ± 2%  -20.21% (p=0.000 n=10)
geomean            491.8n        335.8n       -31.71%

                 │  old.txt   │              new.txt               │
                 │    B/op    │    B/op     vs base                │
Progressbar-8      452.0 ± 0%   420.0 ± 0%   -7.08% (p=0.000 n=10)
ProgressbarInf-8   452.0 ± 0%   356.0 ± 0%  -21.24% (p=0.000 n=10)
geomean            452.0        386.7       -14.45%

                 │  old.txt   │              new.txt               │
                 │ allocs/op  │ allocs/op   vs base                │
Progressbar-8      7.000 ± 0%   4.000 ± 0%  -42.86% (p=0.000 n=10)
ProgressbarInf-8   7.000 ± 0%   5.000 ± 0%  -28.57% (p=0.000 n=10)
geomean            7.000        4.472       -36.11%
```

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [x] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.